refactor: use relative storage compat consumers
